+++ Source/ACE/Network/GameAction/Actions/GameActionQueryHealth.cs
@@ -1,5 +1,6 @@
 ﻿using ACE.Entity;
 using ACE.Managers;
+using ACE.Network.GameAction.QueuedGameActions;
 using ACE.Network.GameEvent.Events;
 using System;
 
@@ -12,7 +13,7 @@ namespace ACE.Network.GameAction.Actions
         {
             uint fullId = message.Payload.ReadUInt32();
 
-            QueuedGameAction action = new QueuedGameAction(fullId, GameActionType.QueryHealth);
+            QueuedGameAction action = new QueuedGameActionQueryHealth(fullId, GameActionType.QueryHealth);
             session.Player.AddToActionQueue(action);
         }
     }
